requirements.txt
pandas
scikit-learn
gensim
numpy
pillow
